requirements.txt
acvl-utils==0.2.5
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
blosc2==3.10.2
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
connected-components-3d==3.25.1
contourpy==1.3.2
cycler==0.12.1
dicom2nifti==2.6.2
dynamic-network-architectures==0.4.2
einops==0.8.1
fft-conv-pytorch==1.2.0
filelock==3.20.0
fire==0.7.1
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
gdcm==1.1
graphviz==0.21
huggingface-hub==0.35.3
idna==3.11
imagecodecs==2025.3.30
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
medmnist==3.0.2
mpmath==1.3.0
msgpack==1.1.2
ndindex==1.10.0
networkx==3.4.2
nibabel==5.3.2
nnunetv2==2.6.2
numexpr==2.14.1
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydicom==3.0.1
pylibjpeg==2.1.0
pylibjpeg-libjpeg==2.3.0
pylibjpeg-openjpeg==2.5.0
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dateutil==2.9.0.post0
python-gdcm==3.2.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shiboken6==6.10.0
simpleitk==2.5.2
six==1.17.0
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.20
torch==2.8.0
torchvision==0.23.0
TotalSegmentator==2.11.0
tqdm==4.67.1
traceback2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
unittest2==1.1.0
urllib3==2.5.0
vtk==9.5.2
xvfbwrapper==0.2.14
yacs==0.1.8
